+++ public/shaders/vorticity.wgsl
@@ -2,7 +2,7 @@
 @import "common.wgsl";
 
 // --- Resource declarations for vorticity calculation ---
-@group(1) @binding(0) var velocityIn: texture_3d<f32>;
+@group(1) @binding(0) var velocityIn: texture_3d<min16float>;
 @group(1) @binding(1) var vorticityOut: texture_storage_3d<rgba16float, write>;
 
 @compute @workgroup_size(4,4,4)
@@ -31,7 +31,7 @@ fn main(@builtin(global_invocation_id) id : vec3<u32>) {
   let dvy_dx = vp_x1.y - vn_x1.y;
   let dvx_dy = vp_y1.x - vn_y1.x;
 
-  let vorticity = vec3<f32>(dvz_dy - dvy_dz, dvx_dz - dvz_dx, dvy_dx - dvx_dy) / (2.0 * params.dx);
+  let vorticity = vec3<min16float>(dvz_dy - dvy_dz, dvx_dz - dvz_dx, dvy_dx - dvx_dy) / (2.0 * params.dx);
 
   textureStore(vorticityOut, id, vec4f(vorticity, 0.0));
 }
